refactor: reveal image_map
