+++ src/assets/image_source/image_map.ts
@@ -1,0 +1,16 @@
+
+const buildImgPath = (imgName: string) => `/images/${imgName}`; 
+export const image_map = {
+  topBanner: buildImgPath('TopBanner.gif'),
+  campudusLogo: buildImgPath('CampudusLogo.png'),
+  githubLogo: buildImgPath('GitHubLogo.png'),
+  githubLogoTransparent: buildImgPath('GitHubLogoTransparent.png'),
+  linkedInLogo: buildImgPath('LinkedInLogo.svg'),
+  lightLogo: buildImgPath('LightLogo.png'),
+  me: buildImgPath('Me.jpg'),
+  codeBackground: buildImgPath('CodeBackground.png'),
+
+
+
+
+}
