+++ src/components/Profile/Profile.tsx
@@ -1,5 +1,5 @@
 
-import { image_map } from "../../resources/image_source/image_map";
+import { image_map } from "../../assets/image_source/image_map";
 const LinkedInURL = "https://www.linkedin.com/in/tsbertram/";
 const Profile = ({ language }: {language: number}) => {
   return (
